FFMpeg movie driver refactor
The actual library code has been put into helper classes, which have been
split into a legacy (0.9.3 forever) version for the arcade build and
a version that compiles with ffmpeg 3.0 for home builds.
