+++ build-home.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
 ./autogen.sh
-#./configure --with-x --with-gnu-ld --target=i386-pc-linux-gnu --host=i386-pc-linux-gnu
-./configure --with-x --with-gnu-ld
+./configure --with-x --with-gnu-ld --target=i386-pc-linux-gnu --host=i386-pc-linux-gnu
+#./configure --with-x --with-gnu-ld
 if [ "x$1" = "x" ];
 then
     echo "Running make"
